backend/requirements.txt
Package                 Version
----------------------- -----------
absl-py                 2.1.0
annotated-types         0.7.0
anyio                   4.8.0
astunparse              1.6.3
certifi                 2024.12.14
charset-normalizer      3.4.1
click                   8.1.8
contourpy               1.3.1
cycler                  0.12.1
fastapi                 0.115.6
flatbuffers             24.12.23
fonttools               4.55.3
gast                    0.6.0
google-pasta            0.2.0
grpcio                  1.69.0
h11                     0.14.0
h5py                    3.12.1
idna                    3.10
keras                   3.8.0
kiwisolver              1.4.8
libclang                18.1.1
Markdown                3.7
markdown-it-py          3.0.0
MarkupSafe              3.0.2
matplotlib              3.10.0
mdurl                   0.1.2
ml-dtypes               0.4.1
namex                   0.0.8
numpy                   2.0.2
opencv-python           4.10.0.84
opt_einsum              3.4.0
optree                  0.13.1
packaging               24.2
pillow                  11.1.0
pip                     24.2
protobuf                5.29.3
pydantic                2.10.5
pydantic_core           2.27.2
Pygments                2.19.1
pyparsing               3.2.1
python-dateutil         2.9.0.post0
python-multipart        0.0.20
requests                2.32.3
rich                    13.9.4
setuptools              75.8.0
six                     1.17.0
sniffio                 1.3.1
starlette               0.41.3
tensorboard             2.18.0
tensorboard-data-server 0.7.2
tensorflow              2.18.0
termcolor               2.5.0
typing_extensions       4.12.2
urllib3                 2.3.0
uvicorn                 0.34.0
Werkzeug                3.1.3
wheel                   0.45.1
wrapt                   1.17.1
